Move foreign toplevel helper classes around
